requirements.txt
# Core dependencies
flask>=2.3.0
werkzeug>=2.3.0
Pillow>=9.0.0
numpy>=1.21.0

# Background removal (lightweight)
rembg>=2.0.50

# Optional heavy dependencies (commented out for size optimization)
# torch>=1.9.0
# torchvision>=0.10.0
# opencv-python>=4.5.0
# scikit-image>=0.19.0
# realesrgan>=0.3.0
# basicsr>=1.4.2
# gfpgan>=1.3.8
